+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.25.0
 
-* `app-vue-vuetify` now supports [TypeScript](https://www.typescriptlang.org/), which is default
+* `app-electron-vuetify` and `app-vue-vuetify` now support [TypeScript](https://www.typescriptlang.org/), which is default
 * removed `app-electron-mdbootstrap` and `app-vue-mdbootstrap`
 
 ## 0.24.0